chore: move statistic files to the admin folder
